requirements.txt
psycopg2-binary>=2.9.9
cx_Oracle>=8.3
numpy>=1.26
